+++ VultisigApp/VultisigApp/Services/VaultDefaultCoinService.swift
@@ -42,7 +42,7 @@ class VaultDefaultCoinService {
                 
                 Task {
                     do{
-                        try await CoinSelectionViewModel().addDiscoveredTokens(nativeToken: coin, to: vault)
+                        try await CoinService.shared.addDiscoveredTokens(nativeToken: coin, to: vault)
                     } catch {
                         print("The coin \(coin.ticker) could not be added. \(error.localizedDescription)")
                     }
