+++ VultisigApp/VultisigApp/Views/Vault/TokenSelectionView.swift
@@ -216,7 +216,7 @@ struct TokenSelectionView: View {
     
     private func saveAssets() {
         Task {
-            await coinViewModel.saveAssets(for: vault)
+            await CoinService.shared.saveAssets(for: vault,selection: coinViewModel.selection)
         }
     }
 }
